refactor lobby
