--- src/app/admin/[slug]/lobby/LobbyOnboarding.module.scss
@@ -1,145 +1,0 @@
-$mobileBreakpoint: 800px;
-
-.onboardingContainer {
-  max-width: 1200px;
-
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  margin-left: auto;
-  margin-right: auto;
-
-  .header {
-   margin: var(--spacing-48);
-   margin-bottom: 0px;
-  }
-
-  .questionsAndAvatarPreview {
-    margin-top: var(--spacing-32);
-    margin-left:  var(--spacing-48);
-    margin-right:  var(--spacing-48);
-    display: flex;
-    flex-flow: row wrap;
-    flex-direction: row;
-    width: 100%;
-    justify-content: space-evenly;
-
-    @media (min-width: #{$mobileBreakpoint}) {
-      flex-grow: 1;
-    }
-    @media (max-width: #{$mobileBreakpoint}) {
-      margin-left: 0px;
-      margin-right: 0px;
-    }
-
-    .questions {
-      margin-top: 20%;
-      width: 50%;
-      order: 0;
-      @media (max-width: #{$mobileBreakpoint}) {
-        margin-top: 5%;
-        order: 1;
-        width: 90%;
-        margin-left:auto;
-        margin-right:auto;
-      }
-
-      .nameInputAndLabel {
-        display: flex;
-        flex-direction: column;
-        width: 300px;
-
-        input {
-          height: 2rem;
-          border-radius: 8px;
-          border: none;
-          padding-left: var(--spacing-24);
-          margin-top: var(--spacing-24);
-          margin-bottom: var(--spacing-24);
-          font-size: 1rem;
-        }
-      }
-
-      .colorPickerAndLabel {
-        display: flex;
-        flex-direction: column;
-
-        .buttonContainer {
-          width: 300px;
-          display: flex;
-          flex-direction: row;
-          justify-content: space-between;
-          margin-top: 1rem;
-          margin-bottom: 1rem;
-
-          button {
-            width: 3rem;
-            height: 1rem;
-            background-color: "blue";
-            border: none;
-            border-radius: 0;
-            margin-right: 1rem;
-          }
-        }
-      }
-    }
-
-    .entranceButtons {
-      display: flex;
-      flex-direction: row;
-      // justify-content: space-between;
-      margin-top: 2rem;
-      margin-bottom: 2rem;
-    }
-
-    .avatarPreviewContainer {
-      overflow: hidden;
-      // min-height:300px;
-      width: 50%;
-      display: flex;
-      flex-direction: column;
-      gap: 1rem;
-      // justify-content: center;
-      align-items: center;
-      margin-top: 10%;
-      position: relative;
-
-      @media (max-width: #{$mobileBreakpoint}) {
-        width: 100%;
-        margin-top: 0;
-      }
-
-      .svgAndVideo {
-        width: 100%;
-        position: relative;
-        margin-left: auto;
-        margin-right: auto;
-        align-items: center;
-        display: flex;
-
-        video {
-          margin-left: auto;
-          margin-right: auto;
-        }
-
-        svg {
-          position: absolute;
-          width: 100%;
-          top: 0px;
-          // left: 0px;
-          // transform: translate(-50%, 0%);
-          left: 50%;
-          // top: 0px;
-          margin-left: auto;
-          margin-right:auto;
-          overflow: visible;
-        }
-      }
-
-      .avatarPreviewAndLabel {
-        text-align: center;
-        width:100%;
-      }
-    }
-  }
-}
