+++ src/components/Lobby/Lobby.module.scss
@@ -1,0 +1,152 @@
+$mobileBreakpoint: 800px;
+
+.lobbyContainer {
+  display: flex;
+  flex-direction: column;
+  position: relative;
+  width: 100vw;
+  height: 100vh;
+}
+
+// .onboardingContainer {
+//   // width: 90%;
+//   max-width: 1200px;
+
+//   height: 100%;
+//   display: flex;
+//   flex-direction: column;
+//   margin-left: auto;
+//   margin-right: auto;
+
+//   .header {
+//     margin-top: 3rem;
+//     margin-bottom: 3rem;
+//     margin-left: 3rem;
+//     margin-right: 3rem;
+//   }
+
+//   .questionsAndAvatarPreview {
+//     margin-left: 3rem;
+//     margin-right: 3rem;
+//     display: flex;
+//     // flex-grow: 1;
+//     flex-flow: row wrap;
+//     flex-direction: row;
+//     // margin: 2rem;
+//     width: 100%;
+//     // padding: 3rem;
+//     // padding-left: 2rem;
+//     // padding-right: 2rem;
+//     justify-content: space-evenly;
+
+//     @media (min-width: #{$mobileBreakpoint}) {
+//       flex-grow: 1;
+//     }
+//     @media (max-width: #{$mobileBreakpoint}) {
+//       margin-left: 0px;
+//       margin-right: 0px;
+//     }
+
+//     .questions {
+//       width: 50%;
+//       order: 0;
+//       @media (max-width: #{$mobileBreakpoint}) {
+//         order: 1;
+//         width: 90%;
+//         margin-left:auto;
+//         margin-right:auto;
+//       }
+
+//       .nameInputAndLabel {
+//         display: flex;
+//         flex-direction: column;
+//         // margin-top: 2rem;
+//         // margin-bottom: 2rem;
+
+//         input {
+//           height: 2rem;
+//           border-radius: 8px;
+//           border: none;
+//           padding-left: 1rem;
+//           margin-top: 1rem;
+//           margin-bottom: 1rem;
+//           font-size: 1rem;
+//         }
+//       }
+
+//       .colorPickerAndLabel {
+//         display: flex;
+//         flex-direction: column;
+
+//         .buttonContainer {
+//           width: 300px;
+//           display: flex;
+//           flex-direction: row;
+//           justify-content: space-between;
+//           margin-top: 1rem;
+//           margin-bottom: 1rem;
+
+//           button {
+//             width: 3rem;
+//             height: 1rem;
+//             background-color: "blue";
+//             border: none;
+//             border-radius: 0;
+//             margin-right: 1rem;
+//           }
+//         }
+//       }
+//     }
+
+//     .entranceButtons {
+//       display: flex;
+//       flex-direction: row;
+//       justify-content: space-between;
+//       margin-top: 2rem;
+//       margin-bottom: 2rem;
+//     }
+
+//     .avatarPreviewContainer {
+//       overflow: hidden;
+//       min-height:300px;
+//       width: 50%;
+//       display: flex;
+//       flex-direction: column;
+//       gap: 1rem;
+//       justify-content: center;
+//       align-items: center;
+//       position: relative;
+
+//       @media (max-width: #{$mobileBreakpoint}) {
+//         width: 100%;
+//       }
+
+//       .svgAndVideo {
+//         height: 200px;
+//         position: relative;
+//         margin-left: auto;
+//         margin-right: auto;
+//         align-items: center;
+//         display: flex;
+
+//         video {
+//           height: 200px;
+//           margin-left: auto;
+//           margin-right: auto;
+//         }
+
+//         svg {
+//           position: absolute;
+//           top: 0px;
+//           margin-left: auto;
+//           margin-right:auto;
+//           overflow: visible;
+//         }
+//       }
+
+//       .avatarPreviewAndLabel {
+//         text-align: center;
+//       }
+//     }
+//   }
+// }
